read test of accelalation and dps

+++ i2c/multi/l3gd20.h
@@ -1,7 +1,7 @@
 #ifndef _L3GD20_H_
 #define _L3GD20_H_
 
-void L3GD20_readData(int *, int );
+void L3GD20_readData(float *, int );
 void L3GD20_init(int );
 
 #endif // _L3GD20_H_
